// syncro/apps/web/src/lib/api/generated/model/sparepartPriceEntryView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */

export interface SparepartPriceEntryView {
  id?: string;
  sparepartId?: string;
  amount?: number;
  currency?: string;
  kursToIdr?: number;
  idrAmount?: number;
  enteredBy?: string;
  enteredByName?: string;
  enteredAt?: string;
}